requirements.txt
Flask
python-dotenv
gunicorn
requests
PyPDF2